Card assets moved from mtc/assets to a resourcepack (Greatly improved memory usage and performance) || Breaking Change!
Made changes to the card loading to store a resourcelocation instead of loading the texture during init. In order for this to have worked I needed to remove some code, and change quite a few files, as well as change the file structure that the mod uses. Card assets are now moved to the resourcepacks folder inside of a resourcepack. (I could also change it so that the entire mtc folder can be placed in the resource packs folder as well, if thats prefered). Because the card assets are now a resource pack, the folders inside of the pack can no longer have any uppercase letters. This, as well as having to move the card assets inside of "assets/is_mtc/textures/cards/" are the breaking changes for users. By default the mod will look for the cards at the same location as where the cfg files were stored (for example, cfg file is in "Trading card game/card_number_one.cfg", the mod will look for the card in "assets/is_mtc/textures/cards/trading card game/card_number_one.png"). If the card has an illustration path element, then that path is used instead (starting from "cards/").